current/requirements.txt
streamlit
requests
langchain
huggingface-hub
sentence-transformers
python-dotenv